Added Gold Text to PlayState so that you can see how much gold you have left

+++ Source/Components/PlayerData.h
@@ -4,6 +4,6 @@ struct PlayerData
 {
 
 	SBaseStats* myStats; // kanske inte borde vara pointer
-	float myGold;
+	short myGold;
 
 };
